Add files via upload
The clustering code and the output file containing all the tweets, their id and the cluster that they belonged to. Because of memory issues, the original dataset is split in half. This is implemented in the code right after line 271 ('if __name__ == '__main__':). This is where the script should be changed if you want to use the code to cluster a different dataset.